+++ lua/pl/permute.lua
@@ -166,11 +166,11 @@ end
 -- @param ...
 -- @see permute.order_iter
 function permute.iter(...)
-  utils.deprecation_warning {
+  utils.raise_deprecation {
     source = "Penlight " .. utils._VERSION,
     message = "function 'iter' was renamed to 'order_iter'",
     version_removed = "2.0.0",
-    version_deprecated = "1.9.2",
+    deprecated_after = "1.9.2",
   }
 
   return permute.order_iter(...)
@@ -180,11 +180,11 @@ end
 -- @param ...
 -- @see permute.order_iter
 function permute.table(...)
-  utils.deprecation_warning {
+  utils.raise_deprecation {
     source = "Penlight " .. utils._VERSION,
     message = "function 'table' was renamed to 'order_table'",
     version_removed = "2.0.0",
-    version_deprecated = "1.9.2",
+    deprecated_after = "1.9.2",
   }
 
   return permute.order_table(...)