+++ CHANGELOG.md
@@ -12,6 +12,8 @@ see [CONTRIBUTING.md](CONTRIBUTING.md#release-instructions-for-a-new-version) fo
    [#360](https://github.com/lunarmodules/Penlight/pull/360)
  - feat: `permute.list_table` generate table with different sets of values
    [#360](https://github.com/lunarmodules/Penlight/pull/360)
+ - feat: Lua 5.4 'warn' compatibility function
+   [#366](https://github.com/lunarmodules/Penlight/pull/366)
  - feat: deprecation functionality `utils.raise_deprecation`
    [#361](https://github.com/lunarmodules/Penlight/pull/361)
  - fix: `dir.rmtree` failed to remove symlinks to directories